+++ Assets/Appodeal/Common/IRewardedVideoAdListener.cs
@@ -5,7 +5,7 @@ namespace AppodealAds.Unity.Common {
 		void onRewardedVideoLoaded(bool precache);
 		void onRewardedVideoFailedToLoad();
 		void onRewardedVideoShown();
-		void onRewardedVideoFinished(int amount, string name);
+        void onRewardedVideoFinished(double amount, string name);
 		void onRewardedVideoClosed(bool finished);
 	}
 }
